Split out settings fragment
