+++ app/src/main/java/com/twilio/video/app/ui/settings/SettingsFragmentModule.java
@@ -1,6 +1,5 @@
 package com.twilio.video.app.ui.settings;
 
-import com.twilio.video.app.ui.settings.SettingsActivity.SettingsFragment;
 import dagger.Binds;
 import dagger.Module;
 import dagger.android.AndroidInjector;
