+++ app/src/main/java/com/twilio/video/app/ui/settings/SettingsFragmentSubcomponent.java
@@ -1,6 +1,5 @@
 package com.twilio.video.app.ui.settings;
 
-import com.twilio.video.app.ui.settings.SettingsActivity.SettingsFragment;
 import dagger.Subcomponent;
 import dagger.android.AndroidInjector;
